+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/content/entity/Moobloom.java
@@ -4,6 +4,7 @@ import house.greenhouse.bovinesandbuttercups.BovinesAndButtercups;
 import house.greenhouse.bovinesandbuttercups.api.CowType;
 import house.greenhouse.bovinesandbuttercups.api.attachment.CowTypeAttachment;
 import house.greenhouse.bovinesandbuttercups.api.cowtype.OffspringConditions;
+import house.greenhouse.bovinesandbuttercups.content.advancements.criterion.BreedCowWithTypeTrigger;
 import house.greenhouse.bovinesandbuttercups.content.advancements.criterion.MutationTrigger;
 import house.greenhouse.bovinesandbuttercups.content.block.entity.CustomFlowerBlockEntity;
 import house.greenhouse.bovinesandbuttercups.content.component.ItemCustomFlower;
@@ -453,14 +454,18 @@ public class Moobloom extends Cow {
 
             if (getLoveCause() != null)
                 MutationTrigger.INSTANCE.trigger(getLoveCause(), this, otherParent, child, (Holder) randomType);
+            BreedCowWithTypeTrigger.INSTANCE.trigger(getLoveCause(), this, otherParent, child, (Holder) randomType);
             return randomType;
         }
 
         child.particlePositions.clear();
 
-        if (!otherParent.getCowType().equals(getCowType()) && getRandom().nextBoolean())
+        if (!otherParent.getCowType().equals(getCowType()) && getRandom().nextBoolean()) {
+            BreedCowWithTypeTrigger.INSTANCE.trigger(getLoveCause(), this, otherParent, child, (Holder<CowType<?>>)(Holder<?>)otherParent.getCowType());
             return otherParent.getCowType();
+        }
 
+        BreedCowWithTypeTrigger.INSTANCE.trigger(getLoveCause(), this, otherParent, child, (Holder<CowType<?>>)(Holder<?>)getCowType());
         return getCowType();
     }