+++ common/src/main/java/house/greenhouse/bovinesandbuttercups/registry/BovinesCriteriaTriggers.java
@@ -1,5 +1,6 @@
 package house.greenhouse.bovinesandbuttercups.registry;
 
+import house.greenhouse.bovinesandbuttercups.content.advancements.criterion.BreedCowWithTypeTrigger;
 import house.greenhouse.bovinesandbuttercups.content.advancements.criterion.LockEffectTrigger;
 import house.greenhouse.bovinesandbuttercups.content.advancements.criterion.MutationTrigger;
 import house.greenhouse.bovinesandbuttercups.content.advancements.criterion.PreventEffectTrigger;
@@ -10,6 +11,7 @@ import net.minecraft.core.registries.BuiltInRegistries;
 public class BovinesCriteriaTriggers {
 
     public static void registerAll(RegistrationCallback<CriterionTrigger<?>> callback) {
+        callback.register(BuiltInRegistries.TRIGGER_TYPES, BreedCowWithTypeTrigger.ID, BreedCowWithTypeTrigger.INSTANCE);
         callback.register(BuiltInRegistries.TRIGGER_TYPES, LockEffectTrigger.ID, LockEffectTrigger.INSTANCE);
         callback.register(BuiltInRegistries.TRIGGER_TYPES, MutationTrigger.ID, MutationTrigger.INSTANCE);
         callback.register(BuiltInRegistries.TRIGGER_TYPES, PreventEffectTrigger.ID, PreventEffectTrigger.INSTANCE);